Move the function prototypes in a separate header
