+++ src/fe-common/irc/fe-irc-channels.c
@@ -31,6 +31,51 @@
 #include "fe-windows.h"
 #include "window-items.h"
 
+int fe_channel_is_opchannel(IRC_SERVER_REC *server, const char *target)
+{
+	const char *statusmsg;
+
+	/* Quick check */
+	if (server == NULL || server->prefix[(int)(unsigned char)*target] == 0)
+		return FALSE;
+
+	statusmsg = g_hash_table_lookup(server->isupport, "statusmsg");
+	if (statusmsg == NULL)
+		statusmsg = "@+";
+
+	return strchr(statusmsg, *target) != NULL;
+}
+
+const char *fe_channel_skip_prefix(IRC_SERVER_REC *server, const char *target)
+{
+	const char *statusmsg;
+
+	/* Quick check */
+	if (server == NULL || server->prefix[(int)(unsigned char)*target] == 0)
+		return target;
+
+	/* Exit early if target doesn't name a channel */
+	if (server_ischannel(SERVER(server), target) == FALSE)
+		return FALSE;
+
+	statusmsg = g_hash_table_lookup(server->isupport, "statusmsg");
+
+	/* Hack: for bahamut 1.4 which sends neither STATUSMSG nor
+	 * WALLCHOPS in 005, accept @#chan and @+#chan (but not +#chan) */
+	if (statusmsg == NULL && *target != '@')
+		return target;
+
+	if (statusmsg == NULL)
+		statusmsg = "@+";
+
+	/* Strip the leading statusmsg prefixes */
+	while (strchr(statusmsg, *target) != NULL) {
+		target++;
+	}
+
+	return target;
+}
+
 static void sig_channel_rejoin(SERVER_REC *server, REJOIN_REC *rec)
 {
 	g_return_if_fail(rec != NULL);
@@ -46,7 +91,7 @@ static void sig_event_forward(SERVER_REC *server, const char *data,
 	char *params, *from, *to;
 
 	params = event_get_params(data, 3, NULL, &from, &to);
-	if (from != NULL && to != NULL && server_ischannel(server, *from) && server_ischannel(server, *to)) {
+	if (from != NULL && to != NULL && server_ischannel(server, from) && server_ischannel(server, to)) {
 		channel = irc_channel_find(server, from);
 		if (channel != NULL && irc_channel_find(server, to) == NULL) {
 			window_bind_add(window_item_window(channel),
